26/03/2025 Fix dei metodi nei controller che non funzionavano, agiunte la parte di PaymentMethod e finita quella di Address.

+++ src/main/java/it/siinfo/springboot2/service/AddressService.java
@@ -95,6 +95,14 @@ public class AddressService {
     }
 
     public void deleteAddressById(Long id) {
+        Optional<Address> optionalAddress = addressRepository.findById(id);
+        if (optionalAddress.isEmpty()) {
+            throw new EntityNotFoundException("Indirizzo inesistente");
+        }
+        Address address = optionalAddress.get();
+        Users user = address.getUser();
+        user.setAddress(null);
+        userRepository.save(user);
         addressRepository.deleteById(id);
     }
